pyproject.toml
[tool.poetry]
name = "whatstranding"
version = "0.1.0"
description = ""
authors = ["kimminki10 <pangshe10@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"
celery = "^5.4.0"
redis = "^5.2.0"
requests = "^2.32.3"
feedparser = "^6.0.11"
flower = "^2.0.1"
python-decouple = "^3.8"
django-redis = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
